app/requirements.txt
chardet
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
Flask
ipdb
ipython
lxml
matplotlib-inline
numexpr
numpy
pandas
plotly
pyarrow
pydantic
pydantic-settings
python-dateutil
pytz
Werkzeug
gunicorn